refactor: [user settings] create separate file for initial settings
